feat: Include working directory in PR comment (#2)
It is common to want to separate your stacks and use something like a
workflow matrix to execute each in parallel. This could cause multiple
comments and it would be hard to know which one is which until you
expand the diff. By adding the directory to the output, this should help
solve that problem.